Common functions and /list command
