+++ scripts/lib/cmd/ranks.lua
@@ -1,0 +1,5 @@
+-- Table of crew ranks for use with `findString()` from `lib.cmd.common`
+return
+{
+	{function (str) return str:find("^[uU]nt") end,  0, "Untrained"},
+}
