+++ scripts/player/cmd/crew.lua
@@ -1,4 +1,7 @@
 if onServer() then
+package.path = package.path .. ";data/scripts/lib/cmd/?.lua"
+require "common"
+professions = require "professions"
 
 function initialize(action, ...)
 	local flag, msg = false, ""
@@ -32,37 +35,13 @@ function addCrew(ship, profession, amount)
 	return false, err
 end
 
-professions =
-{
-	{function (str) return str:find("^[nN]on") end,  0, "None"},
-	{function (str) return str:find("^[eE]ng") end,  1, "Engine"},
-	{function (str) return str:find("^[gG]un") end,  2, "Gunner"},
-	{function (str) return str:find("^[mM]in") end,  3, "Miner"},
-	{function (str) return str:find("^[rR]ep") end,  4, "Repair"},
-	{function (str) return str:find("^[pP]il") end,  5, "Pilot"},
-	{function (str) return str:find("^[sS]ec") end,  6, "Security"},
-	{function (str) return str:find("^[aA]tt") end,  7, "Attacker"},
-	{function (str) return str:find("^[sS]er") end,  8, "Sergeant"},
-	{function (str) return str:find("^[lL]ie") end,  9, "Lieutenant"},
-	{function (str) return str:find("^[cC]om") end, 10, "Commander"},
-	{function (str) return str:find("^[gG]en") end, 11, "General"},
-	{function (str) return str:find("^[cC]ap") end, 12, "Captain"},
-}
-
 -- Identifies script from a string. No, this time only from a string.
 function getProfession(p)
-	local profession = checkTable(professions, p)
+	local profession = findString(professions, p)
 	if profession then
 		return CrewProfession(profession)
 	end
 	return nil, string.format("Could not identify professionType: %s", s)
-end
-
--- Searches table for patterns.
-function checkTable(table, str)
-	for _,item in pairs(table) do
-		if item[1](str) then return item[2] end
-	end
 end
 
 end
